Opraveni bonusu v lekci 3

+++ 03/cviceni-slovniky-cykly.py
@@ -118,7 +118,7 @@ soucet = 0
 for polozka in ingredience:
     cena_str = polozka[-1]
     # Tady by teoreticky sla pouzit i split a vzit jen prvni prvek pred mezerou
-    cena_cislo = int(cena_str.replace(' kč',''))
+    cena_cislo = float(cena_str.replace(' kč',''))
     soucet += cena_cislo
 
 print(math.ceil(soucet))
